+++ src/Compze.Tessaging/Implementation/Transport/Abstractions/TessageTypeTranslator.cs
@@ -7,7 +7,7 @@ static class TessageTypeTranslator
 {
    public static TransportTessageType TransportTessageType(this Type tessageType)
    {
-      if(tessageType.Is<IExactlyOnceTevent>())
+      if(tessageType.Is<IPublisherIdentifyingTevent<IExactlyOnceTevent>>())
          return Abstractions.TransportTessageType.ExactlyOnceTevent;
       if(tessageType.Is<IExactlyOnceTommand>())
          return Abstractions.TransportTessageType.ExactlyOnceTommand;