+++ src/Compze.Tessaging/Implementation/Transport/Client/Internal/ITessagingInboxConnection.cs
@@ -1,3 +1,4 @@
+using Compze.Abstractions.Public;
 using Compze.Abstractions.Tessaging.Public;
 using Compze.Internals.Transport;
 
@@ -6,5 +7,5 @@ namespace Compze.Tessaging.Implementation.Transport.Client.Internal;
 interface ITessagingInboxConnection
 {
     EndpointInformation EndpointInformation { get; }
-    void EnqueueForDelivery(IExactlyOnceTessage tessage);
+    void EnqueueForDelivery(ITessage tessage, TessageId dedupId);
 }